Shrink type and seqno in struct alias from int to u_short

+++ parse.h
@@ -118,8 +118,8 @@ struct member {
  */
 struct alias {
     char *name;				/* alias name */
-    int type; 				/* {USER,HOST,RUNAS,CMND}ALIAS */
-    unsigned int seqno;			/* sequence number */
+    unsigned short type;		/* {USER,HOST,RUNAS,CMND}ALIAS */
+    unsigned short seqno;		/* sequence number */
     struct member_list members;		/* list of alias members */
 };
